VIEW MATTERMOST - refine calculation logic of buildings for a playable state of game
* GL-40 timer

* GL-40 gone

* GL-40 updated economyUI + fixed bugs

* GL-40 logic reimplemented + malus if production/consumption of producs is not balanced

Closes #40

+++ development/project_urban/Assets/Scripts/BuildingData/PreviewUI/ProcessBuilding/ProcessInstance.cs
@@ -7,7 +7,7 @@ public class ProcessInstance : MonoBehaviour
 {
 
     private FoodEconomy foodEconomy;
-    public static event Action<ProductType, float, ResourceType, float> ProduceProduct;
+    public static event Action<ProductType, int, ResourceType, float> ProduceProduct;
 
     // Aktivierungsstatus je Produkt
     private Dictionary<ProductType, bool> productActive = new();
